Track RB allocation in vrb_map{,_UL} on per-symbol basis

+++ openair2/LAYER2/NR_MAC_gNB/gNB_scheduler_dlsch.c
@@ -545,7 +545,7 @@ void nr_simple_dlsch_preprocessor(module_id_t module_id,
 
   AssertFatal(sched_ctrl->pucch_sched_idx >= 0, "no uplink slot for PUCCH found!\n");
 
-  uint8_t *vrb_map = RC.nrmac[module_id]->common_channels[CC_id].vrb_map;
+  uint16_t *vrb_map = RC.nrmac[module_id]->common_channels[CC_id].vrb_map;
   const int current_harq_pid = sched_ctrl->current_harq_pid;
   NR_UE_harq_t *harq = &sched_ctrl->harq_processes[current_harq_pid];
   NR_UE_ret_info_t *retInfo = &sched_ctrl->retInfo[current_harq_pid];
